Rewrote whole vmm to be more efficient and testable

See merge request eitan318/1001_myos!19

+++ src/kernel/utils/math.h
@@ -7,10 +7,10 @@ extern void __attribute__((cdecl))
 div64_32(uint64_t dividend, uint32_t divisor, uint64_t *quotient_out,
          uint32_t *reminder_out);
 
-static inline uint64_t align_down(uint64_t num, uint64_t jump_size) {
+static inline uintptr_t align_down(uintptr_t num, uintptr_t jump_size) {
   return num & ~(jump_size - 1);
 }
 
-static inline uint64_t align_up(uint64_t num, uint64_t jump_size) {
+static inline uintptr_t align_up(uintptr_t num, uintptr_t jump_size) {
   return align_down(num + jump_size - 1, jump_size);
 }
